analyzer/requirements.txt
flask==3.0.3
flask-cors==4.0.1
uvicorn==0.30.1
gunicorn==22.0.0
python-dotenv==1.0.1

# Optional LLM backends (choose one or both depending on your setup)
llama-cpp-python==0.2.86; platform_system == "Linux"
transformers==4.42.4
accelerate==0.33.0
# bitsandbytes is Linux/CUDA only; make it optional via environment marker
bitsandbytes==0.42.0; platform_system == "Linux"
sentencepiece==0.2.0
safetensors==0.4.3

# Utilities
orjson==3.10.7
rapidfuzz==3.9.6

